Fix for missing iterations. The problem occurred when you have a dependent expression and an alignment change during an iteration. The alignment change can only happen when the current alignment is None and the new column has aligned values. The problem caused some iterations to be skipped.
Here is an example rule:

report b123b severity error
	a #= for (x in (1,2,3)) $x;

	b #=  for (y in ($a,)) $y;

	$b >= 2 and exists(Revenues[[]])

This rule should produce twice the number of Revenues facts that are present in the filing.

The problem, was when adding a dependent column, the code was mis-identifying that the alignment whas being switched. This occured when the for body column was added for the value of $x.

The fix was to correctly identify when the alignment switch was happening.

Also, some updates to the --xule-debug-table features.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 21817]