Fix issue with unbound values in that tags during message processing. Example:
output x
@a#a + @b#b
message "{$a} + {$b}

If @b is unbound, the message was producing an error because the tag of $b is unbound and causes a stop iteration during the evaluation of the message. The fix is when the context is copied for message processing, each unbound tag is converted to a 'none'. 

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 22654]