Fix handling of tags on sub tables. Before, the tags were assigned to the values pulled off the sub table, but not they are copied. This fixes an issue where the tags would have the last value from the sub table for all values of the sub table because the tags for each value where pointing to the same object.
Fixed passing down tags when a sub table is created. This fixed an issue when evaluating a message that uses a sub table (i.e. rule-focus with list()).

Added hooks to the xule processor for other plugins to call xule to compile and run rules. This was needed for the FERC/render plugin

Removed check that the rule result can only produce one value. Althought this is still true, have a list() function in the rule-focus caused this to raise an error when it really isn't an error. Now if there is more than one iteration in a result, only the first one will produce a result.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 22838]