Update to handle iterables in filters. This replaces all changes from 23067 to 23073.
Filters that contained something iterable in the where or returns did not work correctly. The fix was to treat iterables in each loop of the filter as separate columns on the iteration table instead of as the same column. For example:

filter list(list(1,2),list(3,4)) returns list(for $x in $item $x+10)

This should return list(list(11,12),list(13,14))

The 'for' expression has a single node_id in the AST. Before this change, there was only 1 column on the iteration table for the 'for' expression. Now each iteration creates a separate column in the iteration table.

Also, fix handling of the $item variable in nested filters.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23074]