Updated handling of properties when the object is a fact and value is a list or set. This happens with setitem extensible enumerations. The fact value is a set of qnames.
When using a fact set, the fact properties did not work correctly because the processor saw the objet as a set and tried to applie the property to each member of the set. So a factset like {@myEnumFact}.period would fail. 

Now, the property evaluator checks if the property is applicicable to a fact even if the fact is a set or list.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23219]