Updated XULE processor for comment from Derek at Workiva posted in Github.  Stop recreating fact index.
Xule creates a fact index for each rule set and the ferc renderer has a rule set for each template, so it's recreating an identical index for each template. Removing this redundancy improves performance by 30-40%.

Xule does add to the index at one point during rule evaluation to copy information for aspects not included within the instance (and so not discovered during the initial indexing stage).

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23358]