Updated Render for comment from Derek at Workiva:
Description
If an invalid date with a {http://www.xbrl.org/inlineXBRL/transformation/2010-04-20}dateslashus transformation is run through the Renderer, it produces the following error:
File "/usr/local/lib/python3.9/site-packages/ferc_renderer/render.py", line 881, in substitute_rule
    fact_number)
  File "/usr/local/lib/python3.9/site-packages/ferc_renderer/render.py", line 2445, in format_fact
    scale)
  File "/usr/local/lib/python3.9/site-packages/ferc_renderer/render.py", line 2671, in format_dateslahus
    return model_fact.xValue.strftime('%m/%d/%Y')
AttributeError: 'str' object has no attribute 'strftime'

This error only appears to be affecting FERC v1.5 at the moment. But to be on the safe side we should go ahead and guard this transformation like some of the other transformations. By raising a FERCRenderException it allows this except to render the supplied value, but adds a log in the log.txt that is included in renderer zip.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23359]