Fixed properties: role, role-description, arcrole, arcrole-description.
These properties translate the role or arcrole uri to the Arelle model object. However, it was only using the modelXbrl in the xule_context which is the same as the model fromthe loaded document. When the role/arcrole was from a taxonomy loaded with the xule taxonomy() function, it did not work correctly. 

The fix pulls the Arelle model from the property object (the object that property is on) instead of the xule_context.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23374]