Back out revision 104 from //ConsistencySuite/Server_Xule_Processor_DEV-Phillip/xule/XuleProperties.py
This contains changes to move the to-xince property out and added the add_property() function to allow dynamic adding of properties. The to-xince will then be added to the Xince plugin.

Also contains a fix to the .concepts() property to not include the item and tuple elements in the results.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23533]