Updated .to-set and to-list to add the current_alignment to the result value. This was causing problem in a rule like:
output p-test3e
  $x = list([@Assets])
  dict(list('lang',  $x.to-list))

output p-test3f
  $x = list([@Assets])
  dict(list('lang',  $x.to-set))

The alignment from the .to-list and to-set was not being carried forward. This caused the dictionary to have a list that was None aligned but the aligned_result_only flag was true so the dict() funtion would skip the value and a return an empty dictionary.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23534]