Added for loop hack for aggregation.
The no-for-hack is a change to how a for loop in an aggregation function (list, set, dict) will work with unbound or skip values. With the hack, which is the default, if there is an unbound or skip value in the collection of values or an alignment, "if will cause the iteration to be skipped. This can happen when filtering a list with using a for:

     list(for $x in list(1,2,3, none) if $x == none skip else $x)

This looks like is should produce a list of 1,2,3, however, this will instead produce no result because the skip is aggregated witht he 1,2,3 and then the whole aggregated iteration is skipped. Instead a filter should be used:

    filter list(1,2,3,none) where $item != none

This will produced the expected result of a list of 1,2,3 (without the none). To produce the list using the for loop inside the aggregation a hack was added to the code to handle this specific case. The hack is on by default. This option will turn off the hack to work in the correct way. THE HACK IS INTENDED TO BE TEMPORARY. THE DEFAULT MAY BE CHANGED TO RUN WITHOUT THE HACK. To future proof your rules, use filters instead of for loops inside aggregation when that is what is intended.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23567]