Fix for @instance in a factset. When processing dimensions in a factset, the code was using the model from the loaded instance/taxonomy which is incorrect when the @instance is use. It uses the taxonomy to find the default member for the dimensions to check if the supplied member is the default.
This corrects the initial issue. However, another issue was discovered. When processing the factset using namespace groups as the dimension aspect. In this case, xule will produce an error stating that namespace group prefixed names cannot be used for a dimension aspect. This greatly reduces the utility of using namespce grouped prefixes. But it brings up issues around how to align dimensions that use a namespace group. For example if a factset uses @us-gaap:Assets where us-gaap is a namespace group for all us gaap taxonomy (all versions), the if there are facts wtih us gaap 2022 and 2021 namespaces, should these be aligned even though they have different namespaces. And if so, if another factset uses @us-gaap-2022, should this align with the namespace grouped facts? It gets a bit complicated.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23595]