Updates to run without a file
These changes were an attempt to run xodel without an input doucment. However, its a bit complicated.

I added a check to the CntlrCmdLine.Utility.Run to check if there is a -f entry file supplied. If so, it does nothing and the CntlrCmdLine.Loaded hook will run Xodel. In this situation, all the CntlrCmdLine.Utility.Run for the other plugins will already have run. If there is not  a -f entry file, then Arelle does not call the CntlrCmdLine.Loaded hook. So, instead in the CntlrCmdLine.Utility.Run for Xodel I explicitly call the CntlrCmdLine.Loaded method to start the xodel process.

The problem is, that Xodel depends on Xule to compile first if the --xule-compile option is supplied. This happens in Xule in the CntlrCmdLine.Utility.Run which is run after Xodel's CntlrCmdLine.Utility.Run. To resolve this, I added a --xodel-compile option that will cause Xodel to explicitly call xule's compile. So when running without a -f entry file and you want to compile, you must use --xodel-compile instead of --xule-compile. Checks are in place to ensure that you don't have both a --xodel-compile and a --xule-compile and also if you don't have a -f entry file then you cannot use --xule-compile.

Also the Serializer needs to import SXM during its CntlrCmdLine.Utility.Run which also runs after Xodel. And so all sorts havoc happens. I've worked around this issue for the serializer by adding a hook to the serializer to initialize the serializer (getting the SXM class) and having Xodel call that hook when it does not have the -f entry file.