Fix dictionary handling with in evaluate_comp. I
It was converting the value to a python dictionary to compare, but when comparing to a set/list of dictionaries, this did not work because the set/list containted the dictionary as a set of tuples (2 items, the first is the key the second is the value).  Now dictionaries are converted to a set of tuples for comparison.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23733]