Fix issue with using coverd and @@. The processor was not identifiying that in this situation the factset is really not covered. That is there is alignment. Here's an example:
output pfilt117
list({covered @instance = $INSTANCE1 @@period.days = 91}).count
rule-suffix if $alignment == none 'FIX' else $alignment['period'].end.year

This fix encountered 2 issues related to this. First in the factset, the none aligned value was not tripping the aligned_results_only flag, so we were getting an extra result (for the none aligned). When the factset does not have alignment, it should not trip the flag, so in the example, the flag was not being tripped, but because of the @@ it should be treated as algined.

The second issue was in calculated the cache key. The calculation looks at if the expression has alignment or not. If it has alignment, then the current alignment is used as part of the cache key. In the example, because the expression was not marked as having alignment, it didn't include alignment in the cache key. This caused the processor to use the first calcualted value for the list() instead of recalcing.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23744]