Workiva changes from pull request 29 to improve cache handling and a mnor imporovement for xule value set.
Instead of using a dictionary for the cache, the cache is now a dictionary like class that allows a maxium byte size. When the cache goes above the max size, it starts deleting items from the cache. This prevents the cache from growing too big.

XuleValueSet now implements __containts__. This provides better performance when the "in" operator is used.

[git-p4: depot-paths = "//ConsistencySuite/Server_Xule_Processor_DEV-Phillip/": change = 23748]