ModelManager updates #29
Xule was change to use the ModelManager that Arelle creates when Arelle is started. This is causing a problem with validation, as the models that xule loads get validated when a user validates in the GUI version of arelle.

These changes create a new ModelManager that is used for taxonomies and instances that are brought in by xule.